Bugfix in DomainList
HUGE bug in DomainList's __contains__... Can't believe it took me this long to
catch it. I had the try: except: inside the loop, rather than outside so cases
like 'google.evil.com' would be considered in the whitelist because when the
lookup to the 2LD 'evil' failed, we'd still be at self.tree['com'] so the lookup
to 'google' would succeed.

Whoops.